# duckdb date filter before recode results in up to 100,000x speed up in spod_get and spod_convert