--- a/inst/extdata/sql-queries/v1-nt-distritos-clean-csv-view-en.sql
+++ b/inst/extdata/sql-queries/v1-nt-distritos-clean-csv-view-en.sql
@@ -1,4 +1,4 @@
-CREATE VIEW nt_csv_clean AS SELECT
+CREATE OR REPLACE VIEW nt_csv_clean AS SELECT
     fecha AS date,
     CAST (CASE distrito
         WHEN 'externo' THEN 'external'