… inst/extdata/sql-queries/v1-nt-distritos-raw-csv-view.sql
… inst/extdata/sql-queries/v1-nt-municipios-clean-csv-view-en.sql
… inst/extdata/sql-queries/v1-nt-municipios-clean-csv-view-es.sql
… inst/extdata/sql-queries/v1-nt-municipios-raw-csv-view.sql
… inst/extdata/sql-queries/v1-od-distritos-clean-csv-view-en.sql
… inst/extdata/sql-queries/v1-od-distritos-clean-csv-view-es.sql
… inst/extdata/sql-queries/v1-od-distritos-raw-csv-view.sql
… inst/extdata/sql-queries/v1-od-municipios-clean-csv-view-en.sql
… inst/extdata/sql-queries/v1-od-municipios-clean-csv-view-es.sql
… inst/extdata/sql-queries/v1-od-municipios-raw-csv-view.sql
… inst/extdata/sql-queries/v2-nt-distritos-clean-csv-view-en.sql
… inst/extdata/sql-queries/v2-nt-distritos-clean-csv-view-es.sql
… inst/extdata/sql-queries/v2-nt-distritos-raw-csv-view.sql
… inst/extdata/sql-queries/v2-nt-gau-clean-csv-view-en.sql
… inst/extdata/sql-queries/v2-nt-gau-clean-csv-view-es.sql
… inst/extdata/sql-queries/v2-nt-gau-raw-csv-view.sql
… inst/extdata/sql-queries/v2-nt-municipios-clean-csv-view-en.sql
… inst/extdata/sql-queries/v2-nt-municipios-clean-csv-view-es.sql
… inst/extdata/sql-queries/v2-nt-municipios-raw-csv-view.sql
… inst/extdata/sql-queries/v2-od-distritos-clean-csv-view-en.sql
… inst/extdata/sql-queries/v2-od-distritos-clean-csv-view-es.sql
… inst/extdata/sql-queries/v2-od-distritos-raw-csv-view.sql
… inst/extdata/sql-queries/v2-od-gau-clean-csv-view-en.sql
… inst/extdata/sql-queries/v2-od-gau-clean-csv-view-es.sql
… inst/extdata/sql-queries/v2-od-gau-raw-csv-view.sql
… inst/extdata/sql-queries/v2-od-municipios-clean-csv-view-en.sql
… inst/extdata/sql-queries/v2-od-municipios-clean-csv-view-es.sql
… inst/extdata/sql-queries/v2-od-municipios-raw-csv-view.sql
… inst/extdata/sql-queries/v2-os-distritos-clean-csv-view-en.sql
… inst/extdata/sql-queries/v2-os-distritos-clean-csv-view-es.sql