@Robinlovelace 2025-06-14
~100,000x speed-ups are vanishingly rare in any walk of life, except software development where I've seen a few such astonishing performance improvements thanks to clever code. Seems like this is a no-brainer.